+++ back-end/CatchAndCast.Service/Services/ProductCharacteristicService.cs
@@ -1,6 +1,7 @@
 ﻿using CatchAndCast.Data.Context;
 using CatchAndCast.Data.Models;
 using CatchAndCast.Service.Dto.Characteristic;
+using CatchAndCast.Service.Exceptions;
 using CatchAndCast.Service.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
@@ -18,6 +19,12 @@ public class ProductCharacteristicService : IProductCharacteristicService
     {
         return await context.Characteristics.Where(x => x.ProductId == id).ToListAsync();
     } 
+
+    public async Task<IEnumerable<ProductCharacteristic>> GetAllAsync()
+    {
+        return await context.Characteristics.ToListAsync();
+    }
+
     public async Task CreateCharacteristicAsync(CreateCharacteristicDto dto)
     {
         var item = new ProductCharacteristic
@@ -33,18 +40,21 @@ public class ProductCharacteristicService : IProductCharacteristicService
     public async Task DeleteAsync(int id)
     {
         var item = await context.Characteristics.FindAsync(id);
+        if (item is null)
+        {
+            throw new ItemNotFound();
+        }
         context.Characteristics.Remove(item);
         await context.SaveChangesAsync();
-    }
-
-    public async Task<IEnumerable<ProductCharacteristic>> GetAllAsync()
-    {
-        return await context.Characteristics.ToListAsync();
     }
 
     public async Task UpdateDescription(UpdateCharacteristicDto dto)
     {
         var item = await context.Characteristics.FindAsync(dto.Id);
+        if (item is null)
+        {
+            throw new ItemNotFound();
+        }
         item.DescriptionOfCharacteristic = dto.Description;
         await context.SaveChangesAsync();
     }
